[追加]帯域情報を追加
